web/app/services/update-countries/requirements.txt
Flask==1.1.2
flask-restful
Flask-request-params
numpy
pandas_gbq
requests
google-oauth
google-oauth2-tool
gunicorn
astropy>=4.0
julian>=0.14
pandas>=1.0
matplotlib>=3.2
scipy
PyAstronomy
ruamel.yaml
ipywidgets
ipython
bokeh
setuptools